// TimeWarden.Client/timewarden-client/src/api/generated/models/invoiceVM.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TimeWarden.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ClientVM } from './clientVM';
import type { InvoiceStatus } from './invoiceStatus';
import type { ItemOfWorkVM } from './itemOfWorkVM';
import type { UserVM } from './userVM';

export interface InvoiceVM {
  /** @nullable */
  id?: string | null;
  clientId?: string;
  invoiceDate?: string;
  status?: InvoiceStatus;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  itemCount?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  totalAmount?: number | string;
  /** @nullable */
  extraNotes?: string | null;
  itemsOfWork?: ItemOfWorkVM[];
  client?: ClientVM;
  user?: null | UserVM;
}